requirements.txt
snowflake-snowpark-python[pandas] @ git+https://github.com/snowflakedb/snowpark-python@dev/local-testing
